dataplane: extract flow-config accessors to maps_flow.go (#1686)
Same-package code motion only; method bodies byte-identical (verified).
Part of the #1686 split of pkg/dataplane/maps.go into domain accessor
files.
